Add automatic end-of-interrupt
With this change, the glue logic will automatically clear the
interrupt latch, meaning software will no longer need to do
a Xosera read to clear it.

In discussion on Discord this was felt to be more useful, and
neatly sidesteps a few issues when a non-Xosera specific handler
is registered (i.e. if level 2 interrupts are enabled without
a handler other than the basic one provided by the firmware).

It also makes races less likely when multiple interrupts overlap
(e.g. another is raised while one is still in-service).